# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "",
        "contact": {
            "email": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://notin.ir/api",
            "description": "Production Server"
        },
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/coin-packages": {
            "get": {
                "tags": [
                    "Coin Packages"
                ],
                "summary": "دریافت لیست پکیج‌های کوین",
                "description": "بازیابی لیست تمام پکیج‌های کوین فعال برای خریداری",
                "operationId": "79400b6d7464a26c67bfed8c1d6ab9a3",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "جستجو بر اساس نام پکیج",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "پکیج 100 کوین"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "لیست پکیج‌ها دریافت شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "پکیج 100 کوین"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "پکیج مقدماتی"
                                                    },
                                                    "image": {
                                                        "description": "لینک فایل امضا شده برای تصویر",
                                                        "type": "string",
                                                        "example": "https://example.com/storage/images/package-1.jpg?signature=xxx"
                                                    },
                                                    "coins": {
                                                        "description": "تعداد کوین",
                                                        "type": "integer",
                                                        "example": 100
                                                    },
                                                    "price": {
                                                        "description": "قیمت اصلی",
                                                        "type": "integer",
                                                        "example": 100000
                                                    },
                                                    "discount_percentage": {
                                                        "description": "درصد تخفیف",
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-01-01T00:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-01-01T00:00:00Z"
                                                    },
                                                    "final_price": {
                                                        "description": "قیمت بعد از تخفیف",
                                                        "type": "integer",
                                                        "example": 90000
                                                    },
                                                    "discount_amount": {
                                                        "description": "مبلغ تخفیف",
                                                        "type": "integer",
                                                        "example": 10000
                                                    },
                                                    "link_cafebazaar": {
                                                        "description": "لینک کافه بازار (اختیاری)",
                                                        "type": "string",
                                                        "example": "https://cafebazaar.ir/..."
                                                    },
                                                    "link_myket": {
                                                        "description": "لینک مایکت (اختیاری)",
                                                        "type": "string",
                                                        "example": "https://myket.ir/..."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "عدم احراز هویت",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/coin-packages/purchase-package": {
            "post": {
                "tags": [
                    "Coin Packages"
                ],
                "summary": "خریداری پکیج کوین",
                "description": "خریداری یک پکیج کوین، تأیید مبلغ، و به‌روزرسانی کیف پول",
                "operationId": "e75087caba7c5937d23d9153e821df77",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "coin_package_id",
                                    "paid_amount",
                                    "pay_reference_code"
                                ],
                                "properties": {
                                    "coin_package_id": {
                                        "description": "شناسه پکیج کوین",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "paid_amount": {
                                        "description": "مبلغ پرداخت شده (باید برابر با final_price باشد)",
                                        "type": "integer",
                                        "example": 90000
                                    },
                                    "pay_reference_code": {
                                        "description": "کد پیگیری پرداخت از درگاه پرداخت",
                                        "type": "string",
                                        "example": "PAY-20230101-ABC123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "پکیج با موفقیت خریداری شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "پکیج با موفقیت خریداری شد."
                                        },
                                        "transaction": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "wallet_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "buy"
                                                },
                                                "coins": {
                                                    "description": "کوین اضافه شده",
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "coins_before": {
                                                    "description": "موجودی قبل از خریداری",
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "coins_after": {
                                                    "description": "موجودی بعد از خریداری",
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "paid_amount": {
                                                    "type": "integer",
                                                    "example": 90000
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "خریداری پکیج کوین"
                                                },
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": null
                                                },
                                                "coin_package_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "reference_code": {
                                                    "type": "string",
                                                    "example": "NOTIN-RC-PACKAGE-20230101-ABCDEF"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-01-01T00:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-01-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "پکیج یافت نشد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "پکیج انتخابی در دسترس نیست"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "خطای اعتبارسنجی",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "مبلغ پرداختی باید دقیقاً برابر با قیمت نهایی پکیج باشد"
                                        },
                                        "required_amount": {
                                            "type": "integer",
                                            "example": 90000
                                        },
                                        "paid_amount": {
                                            "type": "integer",
                                            "example": 85000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "خطای سرور",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "خطا در خرید پکیج"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "",
                "bearerFormat": "",
                "scheme": "bearer"
            }
        }
    },
...
}